APIs/netlify.com/2.16.0/python-aiohttp-client/tests/test_deploy_key_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.deploy_key import DeployKey
from openapi_server.models.error import Error


pytestmark = pytest.mark.asyncio

async def test_create_deploy_key(client):
    """Test case for create_deploy_key

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/api/v1/deploy_keys',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_deploy_key(client):
    """Test case for delete_deploy_key

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/api/v1/deploy_keys/{key_id}'.format(key_id='key_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_deploy_key(client):
    """Test case for get_deploy_key

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v1/deploy_keys/{key_id}'.format(key_id='key_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_deploy_keys(client):
    """Test case for list_deploy_keys

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v1/deploy_keys',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

